8-12 work

+++ uhire-API/src/main/java/com/uhire/rest/repository/TaskStatusRepository.java
@@ -1,13 +1,13 @@
 package com.uhire.rest.repository;
 
-import org.springframework.data.mongodb.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.stereotype.Repository;
 
 import com.uhire.rest.model.lists.TaskStatus;
 
 @Repository
-public interface TaskStatusRepository extends JpaRepository<TaskStatus, Integer>, QuerydslPredicateExecutor<TaskStatus> {
+public interface TaskStatusRepository extends JpaRepository<TaskStatus, Long>, QuerydslPredicateExecutor<TaskStatus> {
 
 	TaskStatus findByName(String string);
 }
